requirements.txt
Django==6.0.1
djangorestframework==3.16.1
mysqlclient==2.2.7
python-dotenv==1.2.1
